Add favicon and metadata
Had to make the RootLayout a server component again for the metadata to work. Added a new provider for the excluded paths logic.
