+++ src/app/components/ExcludedPathsProvider.tsx
@@ -1,0 +1,20 @@
+"use client";
+
+import { usePathname } from "next/navigation";
+import Navbar from "./navbar/navbar";
+import Footer from "./footer/footer";
+
+export default function ExcludedPathsProvider(props: {
+  children: React.ReactNode;
+}) {
+  const pathname = usePathname();
+  const isExcludedPage =
+    pathname === "/checkoutpage" || pathname === "/lastpage";
+  return (
+    <>
+      {!isExcludedPage && <Navbar />}
+      {props.children}
+      {!isExcludedPage && <Footer />}
+    </>
+  );
+}
