+++ src/app/layout.tsx
@@ -1,12 +1,10 @@
-"use client";
-
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./styles/globals.css";
 import Footer from "./components/footer/footer";
 import Navbar from "./components/navbar/navbar";
 import { CartProvider } from "./components/cartlogic";
-import { usePathname } from "next/navigation";
+import ExcludedPathsProvider from "./components/ExcludedPathsProvider";
 
 /*const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -18,24 +16,23 @@ const geistMono = Geist_Mono({
   subsets: ["latin"],
 });*/
 
+export const metadata: Metadata = {
+  title: "deCube",
+  description: "Because everything is better in cube form",
+};
+
 export default function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const pathname = usePathname();
-  const isExcludedPage =
-    pathname === "/checkoutpage" || pathname === "/lastpage";
-
   return (
     <html lang="en">
       <body
       //className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
         <CartProvider>
-          {!isExcludedPage && <Navbar />}
-          {children}
-          {!isExcludedPage && <Footer />}
+          <ExcludedPathsProvider>{children}</ExcludedPathsProvider>
         </CartProvider>
       </body>
     </html>
